requirements.txt
yfinance
pandas
numpy
scipy
python-dotenv
requests
python-telegram-bot
pandas-datareader
